`max_reduced_size` option
Summary:
We compute call-site specific reduced code to determine if it is beneficial to inline for size even if there are many callers.
However, there is no point to keep the code around if it is very large.
Thus, here, we clear it out to save memory.

We still cache it when it is small to speed up downstream inlining / shrinking operations.

Reviewed By: agampe

Differential Revision: D75113700

fbshipit-source-id: b805793e731d2e4ab333e26f852e705b6d5496f2