Options for reachability on styles, attributes
Summary:
Add new option to allow for entry points to be established from application code for styles, by changing how walking works to not immediately slurp up a parent ID as used. Doing so would obscure which styles are directly used from external places, beyond just resolving from a child.

This option turns off emitting a reference for a style parent, so that directly using a child (and not its parent directly) can be distinguished from directly using a parent, but collects the parent style's entries and unions them.

Reviewed By: NTillmann

Differential Revision: D71220704

fbshipit-source-id: 5aa1c7551904d371541bf10d3d10562269b16dab