Added a flag to optionally execute fuzzing source blocks
Summary:
Added a new global config called `enable_fuzzing` which, when true, will run a fuzzing pass on source blocks (to be implemented).

The default value of `enable_fuzzing` is false.

Also added the respective getter methods for this flag in `ConfigFiles.cpp`, as it may need to be used when the actual fuzzing is implemented.

Reviewed By: kobywchan

Differential Revision: D75305757

fbshipit-source-id: 751b0158ab0fc379a4a4d2cb904824b7a80f25a8