Adding test cases to check if non-annotated parameters get removed in the RemoveUnusedArgs pass
Summary:
Added test cases for the change in D75252795 to ensure that the code works as intended.

I added an interface into `RemovedUnusedArgsTest.java` called `AnnotatedParameterInterface`, which contains one method with an annotated parameter, and another method with no annotated parameters. The annotation is custom, and defined as `CustomAnnotation`.

The test PostVerify tests, named `DontRemoveAnnotatedParameter` and `RemoveNonAnnotatedParameter` will check if the pass does not remove the method that has annotated parameters, and the other one checks if it does remove the one without annotated parameters.

Reviewed By: wsanville

Differential Revision: D75257372

fbshipit-source-id: f67701ffb9b9fbc00f1b66e7a92a5c2d2884a25f