Adding a check to prevent methods with annotated parameters from being slated for deletion
Summary:
This diff attempts to introduce a fix for the issue listed in (T207767239)[https://www.internalfb.com/T207767239].

Previously, the `RemoveUnusedArgs.cpp` file was able to remove arguments that had annotations. This was causing a Null Pointer Exception in runtime as seen in (S465578}[https://www.internalfb.com/T207767239].

I added a function in `AnnoUtils.cpp` called `has_any_paramater_annotation` to check if a `DexMethod` has any parameter annotation.  If it does, I do not add to `DexMethod` as a candidate for methods to be changed. This change is on line 584 in `RemoveUnusedArgs.cpp`.

However, this is just one potential fix. The current fix will remove the entire method from consideration, instead of just single parameters of the method. The other solution would be to keep parameters that are annotated (so per-parameter basis rather than per-method basis). Some tests could be run to see if this change creates larger app sizes, to determine which approach is better.

The next step in this, would be to develop test cases to ensure that methods with annotated parameters are indeed not being removed.

Reviewed By: wsanville

Differential Revision: D75252795

fbshipit-source-id: 89c185292f968a3efc795774f5b52313e22ed879