Split return patching logic from param patching logic
Summary:
We have a single piece of logic in `patch_parameters_and_returns` to determine if we need to patch the parameters and the return of a given method. It's also reused for multiple cases. That can be confusing and difficult to simplify.

This change splits it up into two pieces that handles the parameters and return value separately. With that we then adopt the `PatchingCandidate` pattern in `collect_return_candidates`.

Reviewed By: NTillmann

Differential Revision: D75273369

fbshipit-source-id: 998ff1957213feffce86bec9cefcc102513a52b2