Run phase one patching as a fixed point iteartion
Summary:
We have a set of simple patching rules that are local to a given class. We might patch a method parameter if we know that its value flows straight to an annotated field. Things like that. Each one of those simple rules do not collide with one another, but overall the updated state might trigger new patching opportunities for another class. That is what's missing in the current execution flow of the patcher.

We now run the set of simple local rules as a repeatable phase. That is we keep running the phase until nothing new is patched anymore. This should help further eliminate nondeterminism in the checker.

Reviewed By: NTillmann

Differential Revision: D75419789

fbshipit-source-id: 68d27ba77e713b9d6804c6e29a6b7c1ec1f279e6