Added support for hot_no_hot_pred to debug tracking, and removed counting of entry blocks
Summary:
Although `hot_no_hot_pred` violations are tracked, there is no support to tracking and printing them like with the other violations in debug builds. I added the support for this type of violation here.

This involves creating a new enum for the violation handler, adding a function to count all blocks with these violations in a `cfg`, and adding support for debug prints.

In addition, it seems like the code for this type of violation previously counted entry blocks (because if there are no preds, it would count). However, I don't believe it makes sense to count entry blocks into this violation, and so I added an if-check to see if there are any predecessors.

After adding the if-check, the number of violations reported dropped from 751517 to 16547 for the Insert Source Blocks Pass.

Reviewed By: kobywchan

Differential Revision: D75540548

fbshipit-source-id: e2f99757dafb1e388c7dffd1b043876134b9e3da