Fix idom violation fixing
Summary: billjzhang's been running some testing and noticed that we don't resolve violations in entry blocks corrects. Upon investigation, I realized that for entry blocks, the idom block is listed as the block itself. To account for this behavior, we just need to check that the idom block is not equal to the current block.

Reviewed By: danjin250

Differential Revision: D76083857

fbshipit-source-id: 62909efd564f7b8ff159988fab90a3445453c6f0