Added metadata structure for fuzzing algorithm and initialized indegrees
Summary:
To create the algorithm to insert fuzzing data, I need to store some extra information that the algorithm will use. This diff creates the `FuzzingMetadata`, which is stored in an `UnorderedMap` of `Block*` to `FuzzingMetadata`, which will make it easy to look up extra info during the fuzzing.

`FuzzingMetadata` contains two fields, `indegrees` and `insertion_id`, which will be used in the queue when traversing the CFG. This diff also aims to calculate the indegrees of the CFG for a method when `insert_custom_source_blocks` is called.

Since the metadata is a struct, it will make it easier to add more metadata fields if necessary in the future.

Reviewed By: kobywchan

Differential Revision: D76301932

fbshipit-source-id: 123ba9f4a8bba0f8a6420e39f8196826f43252b7