Added test to check indegree calculation of CustomValueInsertHelper
Summary:
Wrote a test to check if the new `FuzzingMetadata` gets the correct number of indegrees calculated when the `CustomValueInsertHelper` is used to traverse the graph.

A new `insert_custom_source_blocks_get_indegrees` method was created to expose the indegrees of the inner helper, as without this function, this functionality cannot be tested.

Reviewed By: kobywchan

Differential Revision: D76295105

fbshipit-source-id: 6a04b75167bf3402f1453009d64f3a3c42531a30