Added hot and cold source block counting metrics in InsertSourceBlocks
Summary:
In an effort to fuzz the compiler, statistics on the number of cold and number of hot source blocks may come in handy. This diff adds a post-pass traversal that gathers the metrics, and adds it to the PassManager.

This diff replaces D76529456, as the original intention of that diff is a different scope than this one.

Reviewed By: kobywchan

Differential Revision: D76542727

fbshipit-source-id: 711ea14ad5beb7835f267c4af1dc6c0e34a0f4cd