StyleAnalysis should take a DexStoresVector and compute code roots
Summary:
The StyleAnalysis helper class will be more useful if it allows the member `ReachableResources` to automatically establish code roots. Pass in the dex stores to make that happen.

`additional_roots` is renamed, to make it a bit more clear - this was meant to be a way to pass in ReachableResourcesPluginRegistry items, akin to https://fburl.com/code/i1n66s92

Reviewed By: ssj933

Differential Revision: D76467805

fbshipit-source-id: 20115b019e1edb35c4032ba9c5d596647b77e269