Refactored topo-traversal into a general helper function
Summary:
This diff aims to refactor the topo-traversal into a more general function that can have a process block function supplied to it to fuzz. It also adds a `TopoTraversalHelper` to contain the data needed to perform fuzzing and the function to process each block.

The reason for this refactor is that this traversal can be used for other fuzzing techniques in the future. This includes generating random data and then topo traversing down the CFG to fix violations, or doing a reverse topo traversal, where the fuzzing data first gets generated at the "leaves" and propagates up to the entry block. This refactor helps set up the structure for techniques like this.

Reviewed By: kobywchan

Differential Revision: D76762707

fbshipit-source-id: fac28fc1758cf776923863b13e069203ad818767