Marking StringBuilderOutlinerPass outlined methods as hot with respect to callsites
Summary:
As part of the second part of the referenced task, this diff sets an outlined method to hot if the callsite is hot.

This diff accomplishes this by going through every new outlined method and checking if at least one callsite is hot, if so, inserts hot source blocks into the outlined method.

Reviewed By: NTillmann

Differential Revision: D76468511

fbshipit-source-id: b3d063fd73293f3688f353aa250c2b85a57442e7