Added new violation to track: Hot All Children Cold
Summary: This diff introduces a new violation to be tracked, Hot All Children Cold. This violation essentially gets counted if the last source block of a basic block is hot, but all the entry blocks of the successors are cold (if at least one is hot, this violation is not counter). The only case where this is not counted is if there is a throw instruction after the last source block, as the throw instruction may cause hot -> cold successors.

Reviewed By: kobywchan

Differential Revision: D76841929

fbshipit-source-id: 5a166e584be4c6753785b5adc62f1aa2a4f5b536