Populate env with java.lang.String for const-string opcodes during enum upcast analysis
Summary:
It seems to me that EnumFixpointIterator is computing the type requirements of non-primitive types that can be stored in registers, and uses this to derive what can be safely optimized and how to perform code rewriting.

However, as shown by the test case this does not seem comprehensive enough to guard against mistakes when rewriting the code. Without the change to EnumUpcastAnalysis.cpp, the code in the test case gets optimized like the following, which does not verify during dex2oat: https://www.internalfb.com/phabricator/paste/view/P1840765346?lines=7-8%2C13%2C21-22%2C14

It should actually be possible to optimize/rewrite the code to remove the enum in the test case, but as far as I can tell the existing analyzer doesn't give `CodeTransformer` enough precision to know if there are multiple defs for a given use of an enum, so it's not really set up to transform in a robust way (which could have been done by introducing a new pred block, rewriting branching, and calling the generated `String_valueOf` method for *one* of the defs - but that would be a large overhaul). Excluding such an enum from optimization I think is an appropriate step.

Reviewed By: thezhangwei

Differential Revision: D76840672

fbshipit-source-id: 35d0783bbfa78ed639182389c99e689fcd20fb8a