Added hot->throw->cold block counting metric
Summary: To better aid in tuning the fuzzing algorithm and to potentially aid in analyzing violations, I would like to know the number (and by extension, the percent) of blocks that go from hot source block -> throw -> cold source block. This diff adds that this metric and refactors the `update_source_block_metric` to handle a struct of data.

Reviewed By: kobywchan

Differential Revision: D76920024

fbshipit-source-id: 55c8ac8627679ed458de968dbe6be85a0864e67f