Changed up hot path to child to be more uniformly probabilistic
Summary:
D76771582 attempts to reflect real world values in fuzzing by having hot blocks -> at least one hot child if there is not a throw after a hot source block. However, that diff checks if all children of predecessors are cold before setting the block to hot, which makes it way more likely that the last child processed is the only hot one.

This diff changes that up by speculatively setting one child to hot if a source is hot, with a uniform probability. This should result in less skewed & more uniform fuzzing data.

Reviewed By: kobywchan

Differential Revision: D77035338

fbshipit-source-id: 9da94a9105113156309e1ca86e6af947a85f1c77