More specific escaping for interfaces in SingleImplAnalyze
Summary:
The previous change D45380346 is very conservative, which may cause size regressions.

Make it more specific, so that if there are const-class opcodes for both an interface and its implementor, escape the interface. This is a very coarse check and leaves a bit to chance, but I'd say it's better than SingleImpl behavior prior to D45380346 with no escaping of const-class usages at all.

Reviewed By: thezhangwei

Differential Revision: D77320346

fbshipit-source-id: 1771d53b0760ff609a9813a5537315be333ac7ed