Added Inter-Method Violation Tracking: Hot Callee All Cold Callers
Summary:
Similar to violations in blocks in individual methods, there are also inconsistencies between methods. There inter-method inconsistencies are also violations of control flow, and this diff adds the first violation to track that deals with inter-method interaction.

The violation, Hot Callee All Cold Callers, essentially happens when a Callee is labeled as hot, but all of its callers are labeled as cold (or uninitialized). This is not supposed to happen, because if a callee was run, this means a caller must have called it somewhere. The way this violation is tracked is by checking the cold entry hit value of the entry source block of a method, if it is hot, it will check every caller. In this check, it will find all the source blocks right before an invoke opcode to the callee method, and check if they are all cold or not. If they are all cold (and there exists a caller that is not a GHOST block), the counter will increase. The Violation Tracker currently traverses through the `CallGraph` with a BFS to check each method and count violations.

However, there is a noticeable slowdown when tracking this violation, because it needs to build a `CallGraph`. Right now, this diff uses the most sparse form of the `CallGraph`, the`single_callee_graph`. The other graph generations will be even slower. This does mean complete information is not captured, but this may be enough for now.

Reviewed By: kobywchan

Differential Revision: D77172623

fbshipit-source-id: 5d0ce61b1d60a668da621c5ec32193a0205cb3a2