Added UnorderedMultiMap wrapper to DeterministicContainers
Summary:
Create a wrapper in `DeterministicContainers` for `std::unordered_multimap`. Added respective tests. Added gmock to the test, allowing for the `testing` namespace to be used. Refactored some of the other tests to use the gmock functions.

Note: I exposed an `unordered_equal_range` function for nondeterministic traversal of a range of values since the fixed iterators cannot traverse.

Reviewed By: xuhdev

Differential Revision: D77559836

fbshipit-source-id: ed998ee64595376cd3539189fe544f48cec25816