M2T3b: Delete attributes from .apk files
Summary: Uses builder methods defined in `Serialize.h` to create a new resource table without the attributes that are to be removed. Created a general method `void ResourcesArscFile::modify_attributes` which accepts a lambda function that the user can define to carry out modification logic. This design decision was motivated by the fact that adding, removing, and modifying attributes all have a lot of overlapping code, and the only thing that changes is what we `add` to the `ResComplexEntryBuilder` struct.

Reviewed By: wsanville

Differential Revision: D77468678

fbshipit-source-id: 5addb9470e2dd584b1a4b283ffb9ac84693130a6