M2T3b: Adding instrumentation test for resource deletion
Summary:
Add instrumentation test for `ResourceValueMergingPass`.

In `resource_value_merging_PreVerify`, we check that the `ResourceValueMergingPass::get_resource_optimization` method is returning the correct optimization techniques.

In `resource_value_merging_PostVerify`, we check that the attributes we expected to be deleted are removed from the resource table.

The non-verify files check that the children of the modified attributes still have the same attributes and correct values.

Reviewed By: ssj933

Differential Revision: D77564223

fbshipit-source-id: cf5845c8aee54f1ceb75eb13a3e3dae015b19f09