Fixed logic issue with hot callee all cold callers
Summary:
The hot callee all cold callers violation is a inter-method violation that tracks if a hot callee method has only cold callers. In the prior iteration of tracking this violation, it would skip tracking the edge between ghost entry nodes and methods. The ghost entry node in a callgraph connects to the entry methods of a scope (often <init> <clinit> methods), and therefore it makes sense to treat the edge between the ghost node and entry methods as hot, rather than skipping them.

The reason why skipping it may cause an issue is for the following edge case:

```
GHOST -----> M2
|            |
|            |
|            |
M1 <----------
```

Technically, methods `M1` and `M2` should have hot entry blocks because they are entry methods. But in the case where `M2`calls `M1`, and the source block inside `M2` before the invoke to `M1` is cold while `M1`'s entry block is hot, the old tracker would count this as a violation, even though `M1` is connected to the ghost node and therefore is allowed to be hot without violation. This diff adds a check if a method is connected to the ghost, if so, it does not count it as a violation.

Reviewed By: agampe

Differential Revision: D77751831

fbshipit-source-id: 5aafc13268d5215a0c80477c274b993eb1b092b8