Utilized outer topo-traversal to implement intermethod fuzzing and produce 0 inter-method violations
Summary:
This diff utilizes the outer topo-traversal to the `InsertSourceBlocks` to generate fuzzing data that does not generate intermethod violations.

After each method is processed in the topo-traversal, a scan over the method will record the source block hit values before an invoke `IRInstruction`. Later methods in the topo-traversal will do a lookup on the map, and check if the entry block should be cold or random. The inner topo-traversal will then take the entry hit and generate the rest of the source block values. This repeats until every method in the callgraph has source block values. The parallel walk fills in the rest of the methods not seen in the topo-travesal.

Reviewed By: kobywchan

Differential Revision: D77671823

fbshipit-source-id: 75dc8d7b2c2305f56c096765b0ddb4fdc4e805bc