Replaced std::unordered_multimap with UnorderedMultiMap wrapper
Summary:
Replaced uses of `std::unordered_multimap` with `UnorderedMultiMap` wrapper. I also changed the iterator to be a const_iterator as the codebase seems to expect it to be a `const_iterator`

There are more uses in the Redex codebase, however replacing them causes issues in code outside of Redex (particularly with inheritance of virtual methods), and therefore I have decided against replacing them for now. Error signals from that can be seen in older versions of this diff.

Reviewed By: NTillmann

Differential Revision: D78045235

fbshipit-source-id: d2fdb9502160f1f5cc0afce135ab8523335d346a