Fix attribute rewriting when a node has no attributes
Summary:
The documented return value of `get_attribute_pointer()` as null when there are zero attributes was causing this logic https://fburl.com/code/h4igotck to incorrectly determine where an attribute should be inserted, in the case of adding to a node with zero attributes.

Change that convention to return where attributes would be, since the `ResXMLTree_attrExt` struct does get filled out with a valid values for `attributeStart`, `attributeSize` even when there are zero attributes. All existing callers in our codebase of `get_attribute_pointer()` are already going to behave properly with this change, since they're accessing `count` properly in loops and will avoid dereferencing when count is zero.

Before this change, empty_view_stub.xml would fail with a confusing error message since the position where the addition is makes no sense, which was correctly flagged by the file size assert. Made the asserts better here too.

Reviewed By: NTillmann

Differential Revision: D78591087

fbshipit-source-id: cceda259d3b04731b6e096a462b8fc42bb4d62b9