don't allocate vals vector when its all default values
Summary:
This saves a ton of memory.

We could consider further "compression" approaches, as I suspect that often most interaction values are still the default value, and only a few are not. This diff makes it easier to do further internal representation changes by fully encapsulation the internal representation of the `vals` array.

This is a behavior-preserving change.

Reviewed By: agampe

Differential Revision: D78625614

fbshipit-source-id: b05316ec4987c9d628c52f0bda913573566e0b5d