Added function and flag to print out all violations after CFG
Summary:
This diff adds a utility function `print_cfg_with_all_violating_blocks` that can be run if `print_all_violations` is true. This function prints out all blocks and source blocks, along with their violations below the CFG.

This makes it easier to see all violations at once, as well as have it so that I can parse it in my python visualizer to display violations visually.

Reviewed By: kobywchan

Differential Revision: D78706742

fbshipit-source-id: ed42b4a74de7ad10b6311fcc70261b3623926372