Fill in source blocks when expanding switches
Summary:
Currently, the `expand_switch` part of the `ReduceSparseSwitches` Pass does not fill in source blocks when creating new if blocks. This is causing a lot of control flow violations in Redex, and this diff aims to add source blocks into those new blocks.

This is achieved by checking if there is a hot source block later down in the branches, and if so, adds a hot synthetic block. Otherwise, it will be set to cold (as this means no other branch will be taken later on, and the condition checks will stop early).

Reviewed By: NTillmann

Differential Revision: D78709806

fbshipit-source-id: 018a15ddb884c40b7cd892ce2e8686c0deb64f4e