M3T3: Removing from children when merging attributes
Summary: When an attribute is to be hoisted up to a parent, the attribute must be removed from all children attributes. This diff introduces a method called `remove_attribute_from_descendent` which takes in a parent `resource_id` and its attributes which are to be added (`attr_map`) and the set of operations which are to be applied on the in-memory graph, and adds removal operations for each direction children to `resource_id` for attributes that are going to be added to it but should be removed from the children.

Reviewed By: wsanville

Differential Revision: D78122017

fbshipit-source-id: 2d40d04ed5447bca3e93a5b1d4bc4c9bad818db3