Do not count ghost blocks as ones with uncovered source blocks
Summary:
While investigating uncovered source block violations, a large issue that came up were that most of the flagged methods had uncovered source blocks in the ghost block. This ghost block is always removed later (and has no instructions anyway), but is counted in the violation metrics, making it hard to debug actual examples of blocks with uncovered source blocks.

This diff makes it so that if a block is a ghost (has ghost edges), the violation metric and printer will not count it in.

Reviewed By: NTillmann

Differential Revision: D78820922

fbshipit-source-id: c223cd24c93b86f97d8dd30aea43d0372c3052e8