M4T1: Get modifications for child after merging chains of styles
Summary: As a refresher, `ResourceValueMergingPass::get_resources_to_merge` returns a `std::vector<std::vector<uint32_t>>` that stores a list of resource ID chains which can be merged into some descendant node. `ResourceValueMergingPass::get_style_merging_modifications` returns all the modifications to be applied to the resource table to correctly merge the hierarchy chains. This is achieved by iterating from the head to the tail of each chain (these chains are in the correct merging order) and gathering all the attributes from these resources that will eventually be merged into the tail of the chain.

Reviewed By: wsanville

Differential Revision: D78447938

fbshipit-source-id: f5d4fd09932c6f01124fa927f6ef75bbc90362c2