M4T1: Add new modification for merging
Summary: Add a new type of modification called `StyleModificationSpec::ModificationType::UPDATE_PARENT_ADD_ATTRIBUTES` to be used for merging styles. When merging a parent style into a child's style, the parent's attributes need to be added to the child's attributes and the child's parent should be updated to the parent's parent.

Reviewed By: ssj933

Differential Revision: D78379562

fbshipit-source-id: 963c1676884b785b791d758724ea8de7a4034f77