M4T1: Finding opportunities for resource merging
Summary:
Introduces `std::vector<std::vector<uint32_t>> ResourceValueMergingPass::get_resources_to_merge` to find opportunities to merge resources that have only one child. Since merges can be chained, merging occurs from top to bottom, and the vector stores the parent resource IDs that are to be merged into the child in the order that the merging should happen. We use a vector (as opposed to a set) to enforce the ordering of modifications.
The method returns a `std::vector<std::vector<uint32_t>>` which stores a list of all the chains which can be merged. For example, if two chains can be merged then an example return value may look like: `{{resource_id#1, resource_id#2, resource_id#3}, {resource_id#9, resource_id#10}}`

Reviewed By: wsanville

Differential Revision: D78296925

fbshipit-source-id: c25598758334980f11068a6f907b71cb39d589c4