Insert source blocks when moving return instruction
Summary:
Consider the following:
```
B1:
  load_param v0
  src_blk 1
  const v1 0
  if_eqz v0 :B3
B2:
  const v1 1
  src_blk 2
B3:
  src_blk 3
  return v1
```
This will be reduced to
```
B1:
  load_param v0
  src_blk 1
  const v1 0
  if_eqz v0 :B3
B2:
  const v1 1
  src_blk 2
  return v1
B3:
  src_blk 3
  return v1
```
However, in the case where Source Block 2 is always hit, the instrumentation will (correctly) record that Source Block 3 is never hit. This will lead to later violations upon reinsertion of profiling data because we'll have Source Block 1 (hot) and Source Block 2 (hot) going to Source Block 3 (cold)

Thus, when we move a return instruction, we need to duplicate the source blocks from the original block with the return instruction.

Reviewed By: NTillmann

Differential Revision: D79133065

fbshipit-source-id: 850e4716d4572f6e62cdbfdcdd5c166618774f39