Fix source block value comparison
Summary:
Turns out that source block equality based on values was broken and wasn't correctly comparing the values.

From what I can tell, it would only distinguish values if the hit count of one source block was 1 and the other was 0. That means that Val(3, 1) and Val(1, 1) are equivalent and Val(1, 0) and Val (1, 1) were equivalent.

Adding an explicit != operator seems to fix this

More info:

After further digging and experimentation, it seems like the problem was that there was an implicit conversion of source block values to bool defined. This conversion converted a value to true unless the `val` field was `NAN`. So thus having something like `val1 != val2` would simplify to `true = true`. The fix is the same, we just add an explicit `!=`.

Reviewed By: jimmycFB

Differential Revision: D79607024

fbshipit-source-id: 41fa8ecc20544967e3c83fc41e234dffade15cfe