Fix source block hashing/equality
Summary:
Previously, source blocks were only being compared by the first source block when chained. This created (at least) one known problem in deduplication of block post-fix.

For example, consider the following:
```
      ; A
      ...
      SB 1
      (if-eqz v0 :D)

      (:C)
      SB 2
      (mul-int v0 v0 v0)
      (add-int v0 v0 v0)
      (invoke-static () "LFooX;.bar:()V")
      SB 3 4
      (move-result v1)
      (goto :E)

      (:D)
      SB 5
      (const v1 1)
      (add-int v0 v0 v0)
      (add-int v0 v0 v0)
      (invoke-static () "LFooX;.bar:()V")
      SB 3 6
      (move-result v1)
      (goto :E)

      (:E)
      SB 7
      (return-void)
```
Under the current (i.e. past) logic, this gets mutated to

```
      ; A
      ...
      SB 1
      (if-eqz v0 :D)

      (:C)
      SB 2
      (mul-int v0 v0 v0)
      (goto :E)

      (:D)
      SB 5
      (const v1 1)
      (add-int v0 v0 v0)
      (goto :E)

      (:E)
      (add-int v0 v0 v0)
      (invoke-static () "LFooX;.bar:()V")
      SB 3 4 6
      (move-result v1)
      SB 7
      (return-void)
```

This results in instrumentation problems because there 3, 4, and 6 will always be hot. However, in some cases, 2 or 5 are cold. This leads to a cold source block directly flowing into a hot source block at insertion time.

Tracing this down, it is because we only consider the first source block in the chain when checking for equality and when hashing.

This diff changes that by making our equality function recursive on the source block chain and by creating a custom hash function that takes the hash of all the ids of the source block chain. This uses a set because as long as the same set of source blocks are present, order does not matter.

Reviewed By: ssj933, agampe

Differential Revision: D79298710

fbshipit-source-id: 3d4683ff121698c55f385bdb710d5d7e12be8761