Refactor resource ID check in pb file asserts for better runtime
Summary:
The previous version of `assert_resource_in_one_files` was slow. When running `ResourceValueMergingPass`, it took between 10-20 minutes. This was because for each resource being modified, we were opening and iterating over every pb resource file.

The method has been refactored to coalesce all resource IDs that are being modified, then iterate over all the pb resource files to count the occurrences of all resource IDs and cross-reference these with the resources being modified to ensure they only appear once. The new version of this method runs the pass in 17 seconds.

Reviewed By: wsanville

Differential Revision: D79850260

fbshipit-source-id: cbc78b0c52eb02d8600fa7c354c7e94c95480551