SGT1: Find opportunities between roots
Summary: Introduce a method `find_inter_graph_hoistings` which returns a list of resource ids where a synthetic parent can be introduced to merge common attributes. An exhaustive (O(2^n)) approach includes finding all ways to group the roots of disjoint style graphs and choosing the group the maximizes the number of attributes that can be hoisted to a synthetic node. Since an exhaustive approach will not scale well for Meta's FOA, we create an approximation. The approximation uses a greedy algorithm that starts by finding the best pair of roots (O(n^2 * m)) based on potential attribute savings, then iteratively expands this combination by adding one root at a time that maximizes the total savings where n is the number of valid roots and m is the average number of attributes per root.

Reviewed By: wsanville

Differential Revision: D79313928

fbshipit-source-id: b8ae81fc45e895cdf3d46b5021b448bd52ce5c89