Gate billjzhang's violation fixes behind a param
Summary: We want to run a test on violation fixes as an attempt to get some idea of the ROI of fixing source block violations. This gates the improvements that billjzhang has made behind a flag. We can enable the flag to go back to the old behavior for testing.

Reviewed By: ssj933

Differential Revision: D80048992

fbshipit-source-id: 959cfb4158d26266b1f0a60b8dac66e7a7d93868