Produce baseline profile guided inlines for speed
Summary:
Introduces a bias to the inliner for calls found in the baseline profile.

This bias is composed of two parts:
1. `shrink_discount` - if a callee shrinks when inlined at a particular callsite(due to known information about its arguments), then we favor inlining it. This bias is calculated as `(inlined_size / original_size)^k` where `k` is some parameter to tune the aggressiveness of this bias.
2. `heat_discount` - if a callee is mainly composed of hot code then we want to inline it. This bias uses the percentage `p` of the callee that is hot `(hot_units / (hot_units + cold_units))` (from D80413562). To manage this bias we use a linear mapping using `d` the discount given to a 100% hot callee and `t` the percentage where we assign a neutral bias (1.0). This mapping maps `p=1` to 0 and `p=t` to 1.

This discount is used in `should_inline_at_call_site`, which typically makes inlining decisions on whether or not the inlined size of the method is larger than simply invoking it. We use this bias by multiplying it with the inlined size (smaller bias encourages more inlining).

Reviewed By: NTillmann

Differential Revision: D79905555

fbshipit-source-id: 393540b3cea10e871e5ace7a9e4bc6a4d874caa2