more fine-grained hot/cold inlining behavior options
Summary:
The old "consider_hot_cold" option was designed to prevent inlining cold code into hot code, and did that well. However, it treated code with (present but marked as) "unknown" hotness information as cold code, even though it might be hot. Such "unknown" hotness tends to arise for frequently changing methods/cfgs where the InsertSourceBlocksPass fails to match information.

In this diff, we replace that boolean options with a more fine-grained "hot_cold_inlining_behavior" enum option. In addition to the old behavior ("none" and "respect-not-cold"), it now offers a new behavior "respect-maybe-hot", which treats code with unknown hotness as hot.

Reviewed By: KyleFung

Differential Revision: D81180848

fbshipit-source-id: fac9ea6c19e8b13008379ae903248d51ee8d0f65