Fix source block comparison/hashing for whole blocks
Summary:
We use an expedited comparison method to compare insns in whole blocks for the sake of block duplication. This comparison function is also wrong for source blocks.

The previous comparison function only compared the ids of the first source blocks. This means that different chains of source blocks along with source blocks with the same id but from different methods would be reported as equal.

This is leading to errors in instrumentation.

Reviewed By: NTillmann

Differential Revision: D81368101

fbshipit-source-id: 1d7c74d48572f6057814951875bb69f6a04ad6cd