Promote "New functionality to prune DexAnnotation elements"
Summary:
Promote {bdc7784b11}.

Promote the D81048659 functionality to redex stable, in a turned off state. I plan on gradually turning this functionality on for FoA, but eagerly turning it on in stable for a few Oculus targets where the size improvement is significant, with buy in from the app owners.

Reviewed By: NTillmann

Differential Revision: D82129440

fbshipit-source-id: 3431e18409ebf6eaee4be272fad3d67683800b82