Gracefully handle blocks with throw edges in ReduceArrayLiteralsPass
Summary:
Without the logic change, test case "aputs_with_goto_and_throw_succs" appears to be incorrectly transformed (v1 not defined in all possible executions: https://www.internalfb.com/phabricator/paste/view/P1950003407?lines=42%2C45).

Per discussion with NTillmann the problem stems from the fact that aput either modifies state *or* throws, which is unlike other IR opcodes which are explicitly split into one that could throw and a pseudo instruction to "write" the effect.

Fortunately, constant propagation has a solution for this: BaseEdgeAwareIRAnalyzer. Use that here, but some other changes need to be made, namely the asserts need to be relaxed because instructions are getting analyzed in a different order when edges (and the last instruction of a pred block) are getting analyzed.

Reviewed By: NTillmann

Differential Revision: D82560787

fbshipit-source-id: fb521829205b477e9b5c646df273670fbe5a41f3