Fix direct method didn't count instance argument in null check removal
Summary:
This pass used to differentiate if we need to -1 for parameter index count based on if a method is virtual method or not. This is correct behavior for virtual and static method. However direct method will also need to -1 because they also have an additional load-param for instance parameter.

This diff switch to is_static check so direct method will follow virtual method's handling strategy to consider load-param for instance parameter. While static method don't need to.

Reviewed By: thezhangwei

Differential Revision: D84732332

fbshipit-source-id: d13104b704d403956e30614af8f943311f759726