Remove unused DexInstruction.h include from DefaultAnnotation.cpp
Summary:
Fixed facebook-unused-include-check warning in DefaultAnnotation.cpp by removing the unused include of DexInstruction.h.

The file only uses DexAnnotationSet, DexEncodedValue, and annotation parsing functions, all of which are provided through AnnoUtils.h (which includes DexAnnotation.h and DexClass.h). DexInstruction.h serves only as an umbrella header and is not needed for this test file.

This resolves task T243011187.

[Session trajectory link](https://www.internalfb.com/intern/devai/devmate/inspector/?id=T243011187-c49c74bd-2675-4a32-b0b9-73a66d1eb8b8)

Reviewed By: wsanville

Differential Revision: D85800273

fbshipit-source-id: b474664ebec29c7541b59d93b3ed55fa57a44251