Fix facebook-unused-include-check in DexStoreUtil.h
Summary:
Fixed facebook-unused-include-check clang-tidy violation by moving the #include "Creators.h" from DexStoreUtil.h to DexStoreUtil.cpp.

The issue was that Creators.h is an umbrella header that was included in the header file but only used in the implementation file. Specifically, ClassCreator from Creators.h is only used in the create_canary() function implementation (line 49 of DexStoreUtil.cpp), not in the header file which only contains function declarations.

This change follows the clang-tidy rule that flags includes that "only serve as an umbrella to other headers" and reduces compilation dependencies by keeping the include only where it's actually needed.

Changes:
- Removed #include "Creators.h" from DexStoreUtil.h (line 10)
- Added #include "Creators.h" to DexStoreUtil.cpp (after line 8, before DexEncoding.h)

[Session trajectory link](https://www.internalfb.com/intern/devai/devmate/inspector/?id=T243011379-43c06dbb-76a2-4862-86de-a3b1d2d8fb5b)

Reviewed By: xuhdev

Differential Revision: D85800896

fbshipit-source-id: 48deaf9a1590a2cf16f088feaf8c6ba938384be4